+++ ArtSpectrum/Contracts/Request/UpdateSaleRequest.cs
@@ -2,7 +2,7 @@
 {
     public class UpdateSaleRequest
     {
-        public decimal? VoucherDiscount { get; set; }
+        public int? VoucherDiscount { get; set; }
         public DateTime? StartTimeSales { get; set; }
         public DateTime? EndTimeSales { get; set; }
         public string? VoucherName { get; set; }
